requirements.txt
click
semantic-kernel
python-dotenv
aiohttp
jsonschema
pyyaml
requests
